Added custom element "text" to CKEDITOR.dtd
The Newsletter2Go editor creates CKEditors for elements with the custom tag "text". We must add this element tag to the dtd here, as we experienced that trailing <br> tags have been removed when initializing and destroying CKEditor instances withing the Newsletter2Go editor.

The reason for this is that the CKEditor tries to replace trailing <br> elements with a non-breaking space inside of block-like elements, but as it does not know that "text" is a block-like element, it only removed the <br> element, but did not add any non-breaking space, resulting in removing trailing <br> elements one by one whenever a CKEditor instance was initialized or destroyed.